Move hasMaterialValue into the util layer
